[refactor] reorganize files into new package structure
