+++ src/main/java/org/example/schedule/common/filter/LoginFilter.java
@@ -1,0 +1,4 @@
+package org.example.schedule.common.filter;
+
+public class LoginFilter {
+}
