+++ src/main/java/org/example/schedule/schedule/dto/ScheduleUpdateRequest.java
@@ -1,4 +1,4 @@
-package org.example.schedule.dto;
+package org.example.schedule.schedule.dto;
 
 import lombok.Getter;
