+++ src/main/java/org/example/schedule/schedule/repository/ScheduleRepository.java
@@ -1,6 +1,6 @@
-package org.example.schedule.respository;
+package org.example.schedule.schedule.repository;
 
-import org.example.schedule.entity.Schedule;
+import org.example.schedule.schedule.entity.Schedule;
 import org.springframework.data.jpa.repository.JpaRepository;
 
 public interface ScheduleRepository extends JpaRepository<Schedule, Long> {
